Initial Rendition
Initial Rendition of proof of concept. 1st Change was in the heat exchanger that goes into the cooler of ice water. The aluminum block heat exchanger did not have enough surface area to change the temperature of the water/coolant to be affective. Switched to a PC radiator for liquid cooling and wrapped it in an enclosure to protect the fins.